Refactor add member modal
